Configure shutdown mode instead of working around issue.
In the latest development version of JPype, the way the
Java VM is shut down is configurable. When we tell JPype to
skip "destroying" the JVM, that seems to solve the issue
of extended shutdown delay, as well as issues that emerged
from working around that.

At least according to the test suite, which was run on
Windows and Linux in order to confirm this.